Add tools for backend fault-tolerance
Add CircuitProxy for wrapping storage and caches in an internal circuit
----------------------

The purpose of circuits is to create separation from the app and its
dependencies, but we were breaking that by creating a hard link between
Faulty and its storage backends. Add CircuitProxy wrappers for storage
and cache so that they use internal circuits to break this hard
dependency.

We use circuits directly, outside the containing scope so that the internal
circuits have their own independent backend and do not create a loop.
This will simply use the Memory storage, so that we have a reliable
internal implementation that we can fall back on.

Add FallbackChain storage for falling back to stable storage
----------------------

Although in many cases it's not bad to default circuits to closed on
failure, we can imagine a cascading faulure scenario where disabling
circuits during storage failure could bring a system down. In that case,
we'd like to temporarily fall back to a stable, reliable storage
mechanism. That is what FallbackChain accomplishes. It will try each
storage in order of priority, and if one fails, it will try the next
one.

Some things like locks and resets are always delegated to all storage
backends.

Since we don't yet have a first-class in-memory storage implementation,
FallbackChain is not yet the default configuration, but if we can add a
circuit maximum to Storage::Memory, this could become a default wrapper
for all non-fault-tolerant storage backends.

Add timeout warnings to the Redis storage backend
----------------------

We don't have control over how the user configures their Redis client or
connection pool, so we do some introspection to check whether the
settings are appropriate for circuit breakers. If the user sets high
timeouts, we warn them on $stderr that they should change their
configuration. This warning can be disabled.

Better Fault-Tolerance Documentation
----------------------

Add individual documentation to the README about each fault tolerance
wrapper, and add better advice about how to use each backend.

Separate AutoWire Wrappers for cache/storage
----------------------

The logic to wrap storage and cache backends was getting too complex to
put in the base Faulty class, so we extract them to their own classes.
This also allows for better API documentaiton for how each one works.

Refactor Delegate Classes
----------------------

Refactor storage and cache wrappers to use Forwardable where appropriate
to remove custom code and make delegation more obvious.