Fix redis bug where concurrent CAS requests could crash
The bug happened when multiple clients opened or closed a circuit
concurrently, exposing a bug with the CAS implementation. Fix this bug,
and also add concurrency tests to expose the issue.

Also add support for jruby and truffleruby. This allows us to test
concurrency with real paralllel threads. To add jruby support to specs,
we need to move non-essential gems to the Gemfile instead of the gemspec since
some aren't compatible with jruby.