Remove Scopes and replace them with Faulty instances
The scope terminology is somewhat ambiguous, and requires users to
create an unusual mental model for how Faulty works. Instead, just allow
calling Faulty.new instead of Faulty::Scope.new. This is more intuitive,
and it means there's no longer a "preferred" way to use Faulty. Both ways
are just as easy to use. We emphasize this in the README by having both
the global and object-oriented examples in the main Configuration
section.