Record failures for cache hits
If the cache was hit, but it is time to refresh, we still run a circuit.
However, if the circuit then fails, it still returns the cached value.
In this scenario, we weren't recording the failure, but instead, simply
returning the cached value without recording success or failure at all.

Fix this so that if the circuit fails, it will always record that
failure even if it returns a cached value.