Fix crash in Redis storage backend if opened_at was missing
If the circuit opened_at key was cleared or expired then the Redis
storage backend would crash. Fix this by setting the opened_at time to
its earliest-possible value.

Also change Redis backend to include opened_at in the transaction for
opening and closing circuits. This should reduce the chance that status
and opened_at get out of sync.