Add granular errors for Elasticsearch patch
With this change, the granular errors normally raised by Elasticsearch
will be preserved. Previously we were discarding the granular error
information raised by the Elasticsearch client and simply raising
an "Error" subclass. Now we raise a Faulty error which is always a
subclass of whichever error was raised by Elasticsearch itself.

To do this, we add support for procs in the new error_mapper option for
circuits. This allows us to individually map the errors returned by
Elasticsearch.

Add Faulty::Deprecation for deprecation notification

DEPRECATIONS:

- The error_module option is deprecated. Patches should use the error_module
  option instead. The option will be removed in 0.9